Renamed UI package to View and refactored GameMapUI into Controller and View
Still work to do on the other remaining UI classes
