+++ SolStandard/Containers/View/PauseMenuUI.cs
@@ -9,7 +9,7 @@ using SolStandard.HUD.Menu.Options.PauseMenu.ConfigMenu;
 using SolStandard.Utility;
 using SolStandard.Utility.Assets;
 
-namespace SolStandard.Containers.UI
+namespace SolStandard.Containers.View
 {
     public class PauseMenuUI : IUserInterface
     {
