+++ SolStandard/Containers/View/StatusUI.cs
@@ -9,7 +9,7 @@ using SolStandard.Utility;
 using SolStandard.Utility.Assets;
 using SolStandard.Utility.Monogame;
 
-namespace SolStandard.Containers.UI
+namespace SolStandard.Containers.View
 {
     public class StatusUI : IUserInterface
     {
