+++ SolStandard/SolStandard.csproj
@@ -57,15 +57,16 @@
     <Compile Include="Containers\Contexts\UnitContextualActionMenuContext.cs" />
     <Compile Include="Containers\Contexts\UnitMovingContext.cs" />
     <Compile Include="Containers\Contexts\UnitTargetingContext.cs" />
+    <Compile Include="Containers\Controller\GameMapController.cs" />
     <Compile Include="Containers\IUserInterface.cs" />
     <Compile Include="Containers\MapContainer.cs" />
     <Compile Include="Containers\Contexts\GameMapContext.cs" />
-    <Compile Include="Containers\UI\BattleUI.cs" />
-    <Compile Include="Containers\UI\MainMenuUI.cs" />
-    <Compile Include="Containers\UI\GameMapUI.cs" />
-    <Compile Include="Containers\UI\PauseMenuUI.cs" />
-    <Compile Include="Containers\UI\StatusUI.cs" />
-    <Compile Include="Containers\UI\SelectMapUI.cs" />
+    <Compile Include="Containers\View\BattleUI.cs" />
+    <Compile Include="Containers\View\MainMenuUI.cs" />
+    <Compile Include="Containers\View\GameMapUI.cs" />
+    <Compile Include="Containers\View\PauseMenuUI.cs" />
+    <Compile Include="Containers\View\StatusUI.cs" />
+    <Compile Include="Containers\View\SelectMapUI.cs" />
     <Compile Include="Entity\GameEntity.cs" />
     <Compile Include="Entity\General\BreakableObstacle.cs" />
     <Compile Include="Entity\General\BuffTile.cs" />
