+++ SolStandard/Utility/Events/PickUpCurrencyEvent.cs
@@ -22,7 +22,7 @@ namespace SolStandard.Utility.Events
             GameContext.ActiveUnit.CurrentGold += currency.Value;
             RemoveItemFromMap();
             AssetManager.CoinSFX.Play();
-            GameMapContext.GameMapUI.GenerateObjectiveWindow();
+            GameMapContext.GameMapController.GenerateObjectiveWindow();
             Complete = true;
         }
