+++ SolStandard/Utility/Events/ToggleOpenEvent.cs
@@ -19,7 +19,7 @@ namespace SolStandard.Utility.Events
             if (!openable.IsOpen)
             {
                 openable.Open();
-                GameMapContext.GameMapUI.GenerateObjectiveWindow();
+                GameMapContext.GameMapController.GenerateObjectiveWindow();
             }
             else
             {
